feat: remove npm i step from upgrade script
Since our projects rely on NVM to manage Node.js versions, we are unable to automate switching these versions
since NVM itself is a series of bash scripts and variables which operate on the user's current $PATH. This
commit simply accepts that reality and adds logging to inform the user that they should run `npm install`
and commit the resulting lockfile. Otherwise, we would be able to upgrade Node.js, switch the Node.js
version, and run `npm install` to automatically update the lockfile.